:sparkles: Add blur effect shader and refactor graphic handling to use Menu class
