+++ src/client/Graphic.hpp
@@ -12,6 +12,7 @@
 
 #include <SFML/Graphics.hpp>
 
+#include "Menu/Menu.hpp"
 #include "userInteractions.hpp"
 
 namespace jetpack::Client {
@@ -22,33 +23,15 @@ namespace jetpack::Client {
 			MENU,
 			TYPE_COUNT
 		};
-		bool _isUserNamePressed = false;
 
 		WindowType _windowType;
-		std::string _username;
 		sf::RenderWindow _window;
 		std::map<unsigned int, std::pair<sf::RectangleShape, sf::Vector2f> > _listPlayers;
 		std::mutex _posMutex;
-		sf::Texture _gameBackgroundTexture;
-		sf::Texture _menuBackgroundTexture;
-		sf::Texture _blurTexture;
-		sf::Font _jetpackFont;
-		sf::Color _menuTextColor;
-		sf::Color _menuButtonColor;
-		sf::Color _menuButtonTextColor;
-		sf::Color _menuUsernameBoxColor;
-		sf::Sprite _gameBackground;
-		sf::Sprite _menuBackground;
-		sf::Sprite _blurSprite;
-		sf::Text _menuCountdown;
-		sf::Text _usernameTextButton;
-		sf::RectangleShape _usernameButton;
-		sf::RectangleShape _usernameField;
-		sf::RectangleShape _usernameBox;
-		sf::Text _usernameBoxTitle;
-		sf::Text _usernameBoxContent;
 		std::function<void(UserInteractions_s)> &_sendUserEvent;
 		std::function<void()> _sendChangeUserName;
+
+		Menu _menu;
 
 		void _handleKeyPressed(const sf::Event &);
 		void _handleMousePressed(const sf::Event &);
@@ -66,9 +49,7 @@ namespace jetpack::Client {
 
 		void setUsername(const std::string &name);
 
-		std::string getUsername() const {
-			return this->_username;
-		}
+		std::string getUsername();
 
 		void analyse();
