+++ src/client/main.cpp
@@ -9,7 +9,7 @@
 #include <string>
 
 #include "Logger.hpp"
-#include "ClientParser.hpp"
+#include "Parser/ClientParser.hpp"
 #include "Exception.hpp"
 #include "Program.hpp"
